api_sleep, api_seek override keyrelease rename files in root better clean up on task end extra stdlib funcs make call_routine & queue_routine more predictable add 'about' app clean up override_ funcs

+++ usr/apps.c
@@ -126,12 +126,12 @@ void _start() {
       sort(content->entries, content->size, sizeof(fs_dir_entry_t), sort_filename);
 
    override_draw(0, -1);
-   override_click((uint32_t)&click, -1);
-   override_release((uint32_t)&release, -1);
-   override_resize((uint32_t)&resize, -1);
-   override_hover((uint32_t)&hover, -1);
-   override_keypress((uint32_t)&keypress, -1);
-   override_mouseout((uint32_t)&mouseout, -1);
+   override_click(&click, -1);
+   override_release(&release, -1);
+   override_resize(&resize, -1);
+   override_hover(&hover, -1);
+   override_keypress(&keypress, -1);
+   override_mouseout(&mouseout, -1);
 
    items = 0;
