+++ usr/prog4.c
@@ -21,7 +21,7 @@ void timer_callback() {
     ui_draw(dialog->ui);
     redraw();
 
-    queue_event((uint32_t)(&timer_callback), 6, NULL);
+    queue_event(&timer_callback, 6, NULL);
 
     x+=5;
     y+=5;
@@ -33,7 +33,7 @@ void click_callback(wo_t *wo, int window) {
     (void)wo;
     (void)window;
     clear();
-    queue_event((uint32_t)(&timer_callback), 6, NULL);
+    queue_event(&timer_callback, 6, NULL);
 }
 
 void _start() {
